+++ 103 + 309/Project/Source/StartupDefaultHandler.c
@@ -1,0 +1,7 @@
+#include "IrqDebug.h"
+
+#if !PROJECT_CFG_IRQ_DEBUG_ENABLE
+void IrqDebug_RecordUnhandledVector(void)
+{
+}
+#endif